[fix] Fixed tests
